♻️ auth related files moved to backend/auth
